refactor: refactor auth for chat integrations
